CPIMC.jl

CPIMC.CPIMC — Module

Common functionality to simulate any physical model.

sweep!, measure!, update!

CPIMC.sweep! — Function

sweep!(m::Model, e::Ensemble, c::Configuration, updates, estimators, steps::Int, sampleEvery::Int, throwAway::Int)

Generate a markov chain of length steps using the Metropolis-Hastings algorithm with the updates given in updates. After throwAway steps have been performed, the observables given in estimators are calculated every sampleEvery steps.

Returns a dictionary containing an UpdateCounter for every function given in updates.

CPIMC.measure! — Function

measure!(m::Model, e::Ensemble, c::Configuration, estimators)

Perform measurements on a Configuration. estimators needs to be a dictionary that contains tuples (::OnlineStat, ::Function).

CPIMC.update! — Function

update!(m::Model, e::Ensemble, c::Configuration, updates)

Obtain the next state of the Markov chain by randomly selecting an element of updates and applying the proposed changes to c with the calculated probability. Returns a tuple containing the index of the chosen function and :accept, :reject or :trivial if the function yielded an empty Step.

CPIMC.print_results — Function

print_results(measurements, e::Ensemble)

Print all measured observables, potentially taking the sign into account.

Model

CPIMC.Model — Type

abstract type Model end

Abstract type for the representation of a physical model for which calculations can be performed. Needs to be implemented system-specifically, in particular the following functions need to be defined: (for now)

the one-particle matrix element, energy(::Model, ::Orbital)

the two-particle matrix element, w(::Model, ::Orbital, ::Orbital, ::Orbital, ::Orbital)

CPIMC.Orbital — Type

Abstract type for single-particle basis states, implementation is required for each model.

Configuration

CPIMC.Configuration — Type

Multi-particle trajectory in imaginary time.

Configuration{T} is a parametric type depending on the single-particle basis T <: Orbital.

Fields

occupations :: Set{T} – orbitals which are occupied initially (τ=0)

kinks :: Kinks{T} – collection of excitations and their imaginary times (kinks)

CPIMC.ImgTime — Type

Type alias for imaginary time. FixedPointNumbers are used since these are stable for == and are thus stable as times in Dict.

CPIMC.Δ — Function

Δ(::ImgTime,::ImgTime)

return the length of the periodic interval between the two ::ImgTimes

the periodic distance of two imaginary times:

`Δ(τ1,τ2) = τ2 - τ1` if `τ2 >= τ1` and
`Δ(τ1,τ2) = 1 - (τ1 - τ2) = 1 + τ2 - τ1` else

CPIMC.Kink — Type

const Kink{T} = Union{T2{T}, T4{T}}

Parametric type representing either a one- or a two-particle scattering event given by types 'T2{T}' and 'T4{T}', respectively. This term is motivated by the geometrical interpretation of a path in Configuration Path Integral Monte Carlo formulation of the partition function in occupation number representation. Discrete occupations are given by horizontal lines of (quasi-)particles occupying the corresponding single-particle states and where transitions between these states are thus indicated by vertical 'kinks' connecting the orbital lines that are part of the transition.

CPIMC.Kinks — Type

const Kinks{T} = Vector{Pair{ImgTime, Kink{T}}}

Structure for storing excitations and their imaginary times (kinks).

Can also be constructed by passing such pairs directly.

Kinks(pairs::Pair{ImgTime,<:Kink{T}}...)

CPIMC.T2 — Type

T2{T}

Parametric type representing a 1-particle excitation by specifying a transition from one state (annihilator) to another (creator). In the occupation number representation, this reads

\[a^{\dagger}_i a_j\]

with a creator orbital i and an annihilator orbital j. The single-particle basis is represented by the type parameter T.

CPIMC.T4 — Type

T4{T}

Parametric type representing a 2-particle excitation by specifying a transition from two state (annihilators) to two other (creator) states. In the occupation number representation, this reads

\[a^{\dagger}_i a^{\dagger}_j a_k a_l\]

with creator orbitals i, j and an annihilator orbitals k, l. The single-particle basis is represented by the type parameter T.

CPIMC.basis — Function

basis(c::Configuration{T})

Return type parameter T of the configuration.

CPIMC.orbs — Function

orbs(::T2)

Return a Tuple of all orbitals that are affected by a T2-kink in the conventional ordering i, j.

orbs(::T4)

Return a Tuple of all orbitals that are affected by a T4-kink in the conventional ordering i, j, k, l. This corresponds to the matrix element $w_{ijkl}$ in the same order.

CPIMC.creators — Function

creators(::T4)

return a set of the two creators which are affected by a T4 kink

CPIMC.annihilators — Function

annihilators(x::T4)

return a set of the two annihilators which are affected by a T4 kink

CPIMC.excite — Function

excite(::Set{T}, ::T4{T})

Apply a kink to a set of basis states, i.e. return a set of basis states where the states specified by the creators of the kink are added and the states specified by the annihilators of the kink are dropped from the given set of basis states. This has the physical meaning of a two-particle scattering event where two (quasi-)particles in a many-body state change the single-particle states they occupy. In occupation number representation this reads

\[a^{\dagger}_i a^{\dagger}_j a_k a_l |\{n\}\rangle\]

for creator orbitals i and j and annihilator orbitals k and l. This function assumes fermionic particle statistics (Pauli principle),

\[a^{\dagger}_i a^{\dagger}_i = a_i a_i = 0\]

i.e. the target (creator) states must not be occupied and the initial (annihilator) states must be occupied in the given set of states.

excite(o::Set{T}, κ::Pair{ImgTime,T4{T}})

Apply a T4-kink to a set of basis states for a pair of a time and a T4-kink. This is useful for iterating over a Kinks-object when calculating the occupations at a specific time.

CPIMC.excite! — Function

excite!(::Set{T}, i::T, j::T, k::T, l::T) where {T}
excite!(o::Set{T}, κ::T4{T})

Apply an excitation in-place to a set of basis states that is given by creating orbitals i, j and annihilating the orbitals k, l.

CPIMC.occupations_at — Function

occupations_at(o::Set{T}, kinks::Kinks{T})

Return the occupied orbitals after applying all kinks to initial occupation.

occupations_at(c::Configuration, τ::ImgTime)

Return the occupied orbitals to the right of τ.

CPIMC.next — Function

next(a, τ)

Return index of the first kink after to the given τ. If there is no kink with ::ImgTime larger than τ, return index of first kink. Throw a DomainError if an empty list is passed as first argument.

This function assumes that the kinks in a are ordered with respect to their imaginary time!

CPIMC.prev — Function

prev(a, τ)

Return the index of the first kink earlier than τ::ImgTime. If there is no such kink the last kink is returned. Throw a DomainError if an empty list is passed as first argument.

This function assumes that the kinks in x are ordered with respect to their imaginary time!

weight calculation

CPIMC.ΔWdiag_element — Function

ΔWdiag_element(::Model, ::Ensemble, ::Configuration, i, j, k, l, τ1, τ2)

Calculate the change in the diagonal interaction many-body matrix element due to a change in the occupations given by creating two orbitals i and j and annihilating two orbitals k, l in the interval (τ1, τ2). This interval may be periodically extended over the bounds (0,1) if τ1 > τ2, i.e. the change in the occupation is considered for (τ1,1] ∪ [0,τ2) in that case. We do not need to evaluate the diagonal interaction between all orbitals in all time-intervalls, but it is sufficient to evaluate the full diagonal interaction with the occupations at the start of the intervall and then consider only contributions of orbitals that are changed by kinks in the intervall.

ΔWdiag_element(::Model, ::Ensemble, ::Configuration, i, j, τ1, τ2)

Calculate the change in the diagonal interaction many-body matrix element due to a change in the occupations given by creating an orbital i and annihilating an orbital j in the interval (τ1, τ2). This interval may be periodically extended over the bounds (0,1) if τ1 > τ2, i.e. the change in the occupation is considered for (τ1,1] ∪ [0,τ2) in that case. We do not need to evaluate the diagonal interaction between all orbitals in all time-intervalls, but it is sufficient to evaluate the full diagonal interaction with the occupations at the start of the intervall and then consider only contributions of orbitals that are changed by kinks in the intervall.

CPIMC.ΔW_diag — Function

ΔW_diag(m::Model, i, j, k, l, occ)

change in the diagonal interaction matrix element due to a change in the occupation occ in a periodic interval (τ1,τ2) where no kinks occur the change in the occupation is assumed to consist in a creation of two orbitals i, j and in the annihlation of two orbitals k, l

ΔW_diag(m::Model, i, j, occ)

change in the diagonal interaction matrix element due to a change in the occupation occ in a periodic interval (τ1,τ2) where no kinks occur the change in the occupation is assumed to consist in a creation of one orbitals i and in the annihlation of one orbitals j

CPIMC.ΔWoffdiag_element — Function

ΔWoffdiag_element(::Model, e::Ensemble, itr, itr)

return the change in the offdiagonal many body matrix element of the interaction operator given by adding the kinks in the first iterable and removing the kinks in the second iterable

both arguments itr are required to be iterables containing kinks

CPIMC.ΔT_element — Function

ΔT_element(::Model, i,j,k,l)

return the change in the kinetic many body matrix element due to creating orbitals i, j and annihilating orbitals k, l

CPIMC.Woffdiag_element — Function

Woffdiag_element(::Model, ::Ensemble, ::Orbital, ::Orbital, ::Orbital, ::Orbital)

Return the offdiagonal many body matrix element of the interaction operator

\[\frac{1}{4} ( w_{ijkl} - w_{ijlk} ) (\pm \langle\{\tilde{n}\}|\{n\}^{ij}_{kl}\rangle)\]

for an excitation given by creating orbitals i,j and annihilating orbitals k, l as given by the Slater-Condon rules.

CPIMC.sign_offdiagonal_product — Function

sign_offdiagonal_product(::Model, ::Configuration)

Return the sign of the product of two-particle terms in the offdiagonal many-body matrix elements. These are given by function wminus, i.e.

\[\langle \{\tilde{n}\} | a^{\dagger}_i a^{\dagger}_j a_k a_l | \{n\} \rangle = \pm ( w_{ijkl} - w_{ijlk} ) \text{ for } \{\tilde{n}\} = \{n\}_{kl}^{ij}\]

the term in the braces may be negative and this function returns the product of the sign of all these contributions $w_{ijkl} - w_{ijlk}$ from all kinks. The sign $\pm$ is determined from the permutation factor of the orbitals i,j,k,l and is not calculated here.

used for the calculation the sign of the weight function

CPIMC.ladder_operator_order_factor — Function

ladder_operator_order_factor(sortedlist::Array{<:Orbital,1})

Returns $1$ or $-1$ depending on the order of all ladder operators as given by a list of orbitals.

ladder_operator_order_factor(ck::Kinks)

Returns $1$ or $-1$ depending on the order of all ladder operators as given by the orbitals that affect each kink in the time-ordering of the kinks and in the conventional ordering i, j, k, l, used in the sign estimator.

CPIMC.wminus — Function

wminus(::Model, i,j,k,l)

return the difference of the two-particle matrix element with the same but the last two indices transposed antisymmetric difference of the two-particle matrix elements:

\[w^-_{ijkl} = w_{ijkl} - w_{ijlk}\]

This is called the antisymmetrized two-particle matrix element. This is an abbreviation for these terms arising in the Slater-Condon rules for the calculation of the many-body matrix elements via the one- and two-particle matrix elements of the underlying single-particle basis.

constructing updates

CPIMC.Step — Type

Represents a change which can be applied to a Configuration.

Fields

drop_orbs – Orbitals which are to be removed from the initial occupation

drop_kinks – pairs of ImgTime and Kinks which are to be removed

add_orbs – Orbitals which are added to the initial occupation

add_kinks – pairs of ImgTime and Kinks which are added

CPIMC.apply_step — Function

apply_step(c::Configuration, step::Step)
apply_step(c::Configuration, steps)

Return the result of applying the changes given by the second argument to a Configuration.

CPIMC.apply_step! — Function

apply_step!(c::Configuration, step::Step)
apply_step!(c::Configuration, steps)

Apply the changes given by the second argument to a Configuration.

Missing docstring.

Missing docstring for add_orbs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_orbs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drop_orbs. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drop_orbs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_kinks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for add_kinks!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_kinks. Check Documenter's build log for details.

Missing docstring.

Missing docstring for remove_kinks!. Check Documenter's build log for details.

CPIMC.UpdateCounter — Type

Objects of this type are used to keep track of the acceptance ratio of a class of updates.

Fields

rejected – number of times updates have been rejected

accepted – number of times updates have been accepted

trivial – number of times the changes proposed by updates of that class have been trivial, e.g. the Step returned by update has been empty

For convenience, an outer constructor with all counters set to zero is provided:

UpdateCounter() = UpdateCounter(0,0,0)

utility functions

CPIMC.time_ordered_orbs — Function

time_ordered_orbs(ck::Kinks{T}) where {T <: Orbital}

Get a list of orbitals that affect each kink in the time-ordering of the kinks and in the conventional ordering i, j, k, l.

CPIMC.kinks_affecting_orbs — Function

 kinks_affecting_orbs(itr::Any, itr::Any)

Return all kinks in the first argument that affect one ore more of the orbitals the second argument. The first itr is expected to contain tuples or at least types on which last() is defined and last(i) is intended to contain a ::T4 for all elements i ∈ itr or at least a type which has fields with the names i, j, k, l.

CPIMC.adjacent_kinks — Function

adjacent_kinks(itr::Any, ::Any)

Return a tuple of the kink in itr that is closest right of the time in the second argument. This expects that methods for the functions next() and prev() are defined for the given argument types. Used for getting a tuple of the neighbouring kinks from some imaginary time.

CPIMC.adjacent_kinks_affecting_orbs — Function

adjacent_kinks_affecting_orbs(::Any, ::Any, ::Any)

Return a tuple of the closest kink to the right and the closest kink to the left of some imaginary time that affect one of the orbitals in some collection. If there are no such kinks, return (nothing,nothing). The first argument is expected to be an iterable which contains pairs as elements, with the first element of each pair containing an imaginary time and the second element of each pair containing a kink. The second argument is expected to be a collection of orbitals, that are to be considered regarding affection with one of the kinks in the first argument. The third argument is expected to be an imaginary time from which the neighbouring kinks affected by any orbitals from the second argument are to be determined.

CPIMC.τ_prev_affecting — Function

 τ_prev_affecting(::Any, ::Any, ::Any)

Return the ImgTime of the closest kink to the left of τ that affects one of the orbitals in os.

CPIMC.τ_next_affecting — Function

 τ_next_affecting(::Any, ::Any, ::Any)

Return the ImgTime of the closest kink to the right of τ that affects one of the orbitals in os.

CPIMC.τ_borders — Function

 τ_borders(::Kinks{T}, orbs, ::ImgTime) where {T <: Orbital}

Return a tuple of the ImgTime of the closest kink to the right and the ImgTime of the closest kink to the left of τ that affect one of the orbitals in orbs. If no orbital in os is affected by and kink from the collection in the first argument, return a tuple of the interval bounds (ImgTime(0), ImgTime(1)).

Missing docstring.

Missing docstring for prev. Check Documenter's build log for details.

Missing docstring.

Missing docstring for next. Check Documenter's build log for details.

CPIMC.prev_affecting — Function

 prev_affecting(a::Any, orbs::Any, τ::Any)

Return the index of the closest kink to the left of τ that affects one of the orbitals in orbs. If no orbital in orbs is affected by any kink return 0.

CPIMC.next_affecting — Function

 next_affecting(a::Any, orbs::Any, τ::Any)

Return the index of the closest kink to the right of τ that affects one of the orbitals in orbs. If no orbital in orbs is affected by any kink return 0.

CPIMC.isunaffected — Function

isunaffected(Kinks, orb)

Return if an orbital is not affected by any kink.

CPIMC.isunaffected_in_interval — Function

isunaffected_in_interval(kinks, orb, τ_first::ImgTime, τ_last::ImgTime)

Return if an orbital is not affected by any of the kinks from ck in the open interval (τ_first,τ_last).