Estimators

This module provides a collection of estimators which are independent of the physical system.

CPIMC.Estimators — Module

generic estimators for the CPIMC formalism

CPIMC.Estimators.E — Method

E(m::Model, e::Ensemble, c::Configuration)

estimator for the interaction energy This estimator is redundant because we can calculate the full energy as the sum of the kinetic and the interaction part.

CPIMC.Estimators.Ekin — Method

Ekin(m::Model, e::Ensemble, c::Configuration)

estimator for the kinetic energy

CPIMC.Estimators.K — Method

K(m::Model, e::Ensemble, c::Configuration)

estimator for the number of kinks

CPIMC.Estimators.W — Method

W(m::Model, e::Ensemble, c::Configuration)

estimator for the interaction energy This estimator is redundant because we can calculate the interaction energy from Kfermion and Wdiag.

CPIMC.Estimators.W_diag — Method

W_diag(m::Model, e::Ensemble, c::Configuration)

estimator for the diagonal contribution to the interaction energy

CPIMC.Estimators.W_off_diag — Method

W_off_diag(m::Model, e::Ensemble, c::Configuration)

estimator for the offdiagonal contribution to the interaction energy This estimator is redundant because we can calculate the offdiagonal interaction energy from K_fermion.

CPIMC.Estimators.occupations — Function

occupations(m::Model, e::Ensemble, c::Configuration, emax::Int=100) :: Array{Float64,1}

estimator for the occupations of the emax:Int lowest single particle energy eigenvalues

CPIMC.signum — Method

signum(m::Model, e::Ensemble, c::Configuration)

estimator for the sign of the weight function