PlaneWaves

CPIMC.PlaneWaves — Module

Plane wave single-particle basis states.

CPIMC.PlaneWaves.PlaneWave — Type

struct PlaneWave{D} <: Orbital

Representation of plane wave single-particle states with dimensionality D.

Fields

vec :: StaticVector{D,Int} – momentum vector

spin :: Spin – spin

CPIMC.PlaneWaves.dimension — Function

dimension(o::PlaneWave{D}) where {D}

return the dimension of the momentum vector of an Orbital

dimension(os::Set{<:PlaneWave{D}}) where {D}

return the dimension of the momentum vectors of a Set{<:Orbital{D}}

CPIMC.PlaneWaves.Spin — Type

Type for storing spin projection of particles with spin 1/2.

CPIMC.PlaneWaves.flip — Function

flip(s::Spin)

return opposite spin projection

flip(o::PlaneWave)

return an PlaneWave with the same momentum vector but opposite spin projection as the given o::PlaneWave

CPIMC.PlaneWaves.fractional_spin_polarization — Function

fractional_spin_polarization(occ::Set{PlaneWave)

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

fractional_spin_polarization(c::Configuration{<:PlaneWave})

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

CPIMC.PlaneWaves.ξ — Function

ξ(occ::Set{PlaneWave})

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

ξ(c::Configuration{<:PlaneWave})

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

CPIMC.PlaneWaves.shell — Function

shell(o::PlaneWave{1};dw::Int=2)
shell(o::PlaneWave{2};dw::Int=2)
shell(o::PlaneWave{3};dw::Int=2)

Returns a set of all PlaneWave orbitals lying in the dwth shell.

CPIMC.PlaneWaves.sphere — Function

return two spheres of each spin with radius dk around the wavevector of the argument orbital o

Missing docstring.

Missing docstring for sphere_with_same_spin. Check Documenter's build log for details.