UniformElectronGas

CPIMC.UniformElectronGas — Module

Types and definitions used for simulations of the uniform electron gas.

Model

CPIMC.UniformElectronGas.UEG — Type

Singleton for dispatching.

For other models this would be the place to store additional parameters that might come up in the matrix elements.

CPIMC.w — Method

w(i::Orbital, j::Orbital, k::Orbital, l::Orbital)

Return the two-particle matrix element of the Coulomb interaction. Zero is returned if momentum or spin is not conserved, in consequence of the Bloch-theorem and the spin kronecker-delta in the plane-spin-wave basis.

The singular contribution (i.vec == k.vec) is also removed, i.e. set to zero. This property is not a result of the Bloch theorem for the two-particle Coulomb matrix element in the plane wave basis but is added here to allow for straightforward summation over all combination of orbitals without explictly excluding this component.

It is not part of the UEG many-body Hamiltonian due to the physical assumption that contributions from the uniform background cancel with this diverging term in the thermodynamic limit.

CPIMC.energy — Method

energy(m::UEG, o::PlaneWave)

The single-particle energy of a plane wave with momentum $\vec{k}$ is given by (Hartree atomic units):

$ \epsilon_k = \vec{k}^2 / 2. $

In the internal unit system we use integer $k$-values, which results in:

$ \epsilon_k = \vec{k}^2. $

Units

CPIMC.UniformElectronGas.λ — Function

λ(N, rs, d::Int)

calculate λ, which is the coupling constant of the Hamiltonian in internal units from the particle number N, Bruecker parameter rs for the density and spatial dimension d the differences for d ∈ {2,3} reflect the different expressions for the Coulomb interaction, whichs Fourier component is vq = rac{4π}{q^2} in 3D and vq = rac{2π}{q} in 2D the one-dimensional case involves a short-distance cutoff to avoid the non-integrable divergence of the interaction at zero distance and is not implemented

CPIMC.UniformElectronGas.rs — Function

rs(N, λ, d::Int)

calculate the Brueckner parameter rs for the density from the particle number N and λ, which is the coupling constant of the Hamiltonian in internal units

CPIMC.UniformElectronGas.internal_energy_factor — Function

internal_energy_factor(N, rs, d)

This factor is used to convert an energy quantity from Hartree a.u. to internal units. A value of the quantity in Hartree a.u. has to be multiplied by this factor to yield the corresponding value of the quantity in internal units. A value of the quantity in internal units has to be divided by this factor to yield the corresponding value of the quantity in Hartree a.u.

CPIMC.UniformElectronGas.kF — Function

kF(rs, ξ, d::Int)

calculate the Fermi wavenumber in Hartree a.u. from Brueckner parameter rs for the density, the fractional spin polarization ξ and spatial dimension d

CPIMC.UniformElectronGas.EF — Function

EF(rs, ξ, d::Int)

calculate the Fermi energy in Hartree a.u. from Brueckner parameter rs for the density, the fractional spin polarization ξ and spatial dimension d

CPIMC.UniformElectronGas.β_Ha — Function

β_Ha(Θ, EF)

calculate the inverse temperature in Hartree a.u. from the reduced temperature Θ and the Fermi energy EF in Hartree a.u. the inverse temperature is defined by β = 1 / (kB*T) where kB is the Boltzmann constant and T is the temperature

CPIMC.UniformElectronGas.β — Function

β(Θ, rs, N, ξ, d::Int)

calculate β in internal units