CPIMC

Base.getindex — Method

Base.getindex(kinks::Kinks{T}, τ::ImgTime) where {T}

Allow the kinks[τ]-syntax to retrieve a Kink at a specific ImgTime.

Base.haskey — Method

Base.haskey(ck::Kinks, key::ImgTime)

Check if a Kinks-object contains a kink at a specific time.

CPIMC.apply_step! — Method

apply_step!(c::Configuration, step::Step)
apply_step!(c::Configuration, steps)

Apply the changes given by the second argument to a Configuration.

CPIMC.apply_step — Method

apply_step(c::Configuration, step::Step)
apply_step(c::Configuration, steps)

Return the result of applying the changes given by the second argument to a Configuration.

CPIMC.excitations — Method

excitations(ck::Kinks)

Return a list of the excitations of a Kinks-object, used to allow the use of dictionary syntax.

CPIMC.excite! — Method

excite!(::Set{T}, i::T, j::T, k::T, l::T) where {T}
excite!(o::Set{T}, κ::T4{T})

Apply an excitation in-place to a set of basis states that is given by creating orbitals i, j and annihilating the orbitals k, l.

CPIMC.excite — Method

excite(o::Set{T}, κ::Pair{ImgTime,T4{T}})

Apply a T4-kink to a set of basis states for a pair of a time and a T4-kink. This is useful for iterating over a Kinks-object when calculating the occupations at a specific time.

CPIMC.excite — Method

excite(::Set{T}, ::T4{T})

Apply a kink to a set of basis states, i.e. return a set of basis states where the states specified by the creators of the kink are added and the states specified by the annihilators of the kink are dropped from the given set of basis states. This has the physical meaning of a two-particle scattering event where two (quasi-)particles in a many-body state change the single-particle states they occupy. In occupation number representation this reads

\[a^{\dagger}_i a^{\dagger}_j a_k a_l |\{n\}\rangle\]

for creator orbitals i and j and annihilator orbitals k and l. This function assumes fermionic particle statistics (Pauli principle),

\[a^{\dagger}_i a^{\dagger}_i = a_i a_i = 0\]

i.e. the target (creator) states must not be occupied and the initial (annihilator) states must be occupied in the given set of states.

CPIMC.measure! — Method

measure!(m::Model, e::Ensemble, c::Configuration, estimators)

Perform measurements on a Configuration. estimators needs to be a dictionary that contains tuples (::OnlineStat, ::Function).

CPIMC.print_rates — Method

print_rates(dict::Dict{Any,UpdateCounter})

Calculate and print acceptance statistics.

CPIMC.print_results — Method

print_results(measurements, e::Ensemble)

Print all measured observables, potentially taking the sign into account.

CPIMC.sweep! — Method

sweep!(m::Model, e::Ensemble, c::Configuration, updates, estimators, steps::Int, sampleEvery::Int, throwAway::Int)

Generate a markov chain of length steps using the Metropolis-Hastings algorithm with the updates given in updates. After throwAway steps have been performed, the observables given in estimators are calculated every sampleEvery steps.

Returns a dictionary containing an UpdateCounter for every function given in updates.

CPIMC.times — Method

times(ck::Kinks)

Return a list of the imaginary-times of a Kinks-object, used to allow the use of dictionary syntax.

CPIMC.update! — Method

update!(m::Model, e::Ensemble, c::Configuration, updates)

Obtain the next state of the Markov chain by randomly selecting an element of updates and applying the proposed changes to c with the calculated probability. Returns a tuple containing the index of the chosen function and :accept, :reject or :trivial if the function yielded an empty Step.

Estimators

CPIMC.Estimators.E — Method

E(m::Model, e::Ensemble, c::Configuration)

estimator for the interaction energy This estimator is redundant because we can calculate the full energy as the sum of the kinetic and the interaction part.

CPIMC.Estimators.Ekin — Method

Ekin(m::Model, e::Ensemble, c::Configuration)

estimator for the kinetic energy

CPIMC.Estimators.K — Method

K(m::Model, e::Ensemble, c::Configuration)

estimator for the number of kinks

CPIMC.Estimators.W — Method

W(m::Model, e::Ensemble, c::Configuration)

estimator for the interaction energy This estimator is redundant because we can calculate the interaction energy from Kfermion and Wdiag.

CPIMC.Estimators.W_diag — Method

W_diag(m::Model, e::Ensemble, c::Configuration)

estimator for the diagonal contribution to the interaction energy

CPIMC.Estimators.W_off_diag — Method

W_off_diag(m::Model, e::Ensemble, c::Configuration)

estimator for the offdiagonal contribution to the interaction energy This estimator is redundant because we can calculate the offdiagonal interaction energy from K_fermion.

CPIMC.Estimators.occupations — Function

occupations(m::Model, e::Ensemble, c::Configuration, emax::Int=100) :: Array{Float64,1}

estimator for the occupations of the emax:Int lowest single particle energy eigenvalues

CPIMC.signum — Method

signum(m::Model, e::Ensemble, c::Configuration)

estimator for the sign of the weight function

PlaneWaves

CPIMC.PlaneWaves.dimension — Method

dimension(o::PlaneWave{D}) where {D}

return the dimension of the momentum vector of an Orbital

CPIMC.PlaneWaves.dimension — Method

dimension(os::Set{<:PlaneWave{D}}) where {D}

return the dimension of the momentum vectors of a Set{<:Orbital{D}}

CPIMC.PlaneWaves.find_fourth_orb_for_kink — Method

find_fourth_orb_for_kink(same_kind_ladder_operator, other_kind_ladder_operator1, other_kind_ladder_operator2)

Find the fourth orb to build a Kink accoring to spin/momentum conservation, where the Operator thats acts on the fourth orb is of the same kind (creator/ annihilator) as the one that acts on samekindladderoperator, While on otherkindladderoperator1 and otherkindladder_operator2 the other kind of operators will act.

CPIMC.PlaneWaves.flip — Method

flip(o::PlaneWave)

return an PlaneWave with the same momentum vector but opposite spin projection as the given o::PlaneWave

CPIMC.PlaneWaves.flip — Method

flip(s::Spin)

return opposite spin projection

CPIMC.PlaneWaves.fractional_spin_polarization — Method

fractional_spin_polarization(c::Configuration{<:PlaneWave})

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

CPIMC.PlaneWaves.fractional_spin_polarization — Method

fractional_spin_polarization(occ::Set{PlaneWave)

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

CPIMC.PlaneWaves.sphere — Method

sphere(o::PlaneWave{1}; dk=2)
sphere(o::PlaneWave{2}; dk=2)
sphere(o::PlaneWave{3}; dk=2)

Return the set of all PlaneWaves in a sphere of radius dk around o in momentum space.

CPIMC.PlaneWaves.sphere_with_same_spin — Method

sphere_with_same_spin(o::PlaneWave{1}; dk=2)
sphere_with_same_spin(o::PlaneWave{2}; dk=2)
sphere_with_same_spin(o::PlaneWave{3}; dk=2)

Return the set of all PlaneWaves in a sphere of radius dk around o in momentum space that also have the same spin as o.

CPIMC.PlaneWaves.ξ — Method

ξ(c::Configuration{<:PlaneWave})

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

CPIMC.PlaneWaves.ξ — Method

ξ(occ::Set{PlaneWave})

calculate the fractional spin polarization from a set of Orbitals which each must have a field spin of type @enum Spin Down Up, the fractional spin polarization is defined as the ratio of the absolute difference of the number of particles which Spin Down and Spin Up and the particle number, thus here no convention is made as to which spin component should be occupied more often, as opposed to some literature where N↑ > N↓ is used

UniformElectronGas

CPIMC.UniformElectronGas.EF — Method

EF(rs, ξ, d::Int)

calculate the Fermi energy in Hartree a.u. from Brueckner parameter rs for the density, the fractional spin polarization ξ and spatial dimension d

CPIMC.UniformElectronGas.kF — Method

kF(rs, ξ, d::Int)

calculate the Fermi wavenumber in Hartree a.u. from Brueckner parameter rs for the density, the fractional spin polarization ξ and spatial dimension d

CPIMC.UniformElectronGas.rs — Method

rs(N, λ, d::Int)

calculate the Brueckner parameter rs for the density from the particle number N and λ, which is the coupling constant of the Hamiltonian in internal units

CPIMC.UniformElectronGas.β — Method

β(Θ, rs, N, ξ, d::Int)

calculate β in internal units

CPIMC.UniformElectronGas.β_Ha — Method

β_Ha(Θ, EF)

calculate the inverse temperature in Hartree a.u. from the reduced temperature Θ and the Fermi energy EF in Hartree a.u. the inverse temperature is defined by β = 1 / (kB*T) where kB is the Boltzmann constant and T is the temperature

CPIMC.UniformElectronGas.λ — Method

λ(N, rs, d::Int)

calculate λ, which is the coupling constant of the Hamiltonian in internal units from the particle number N, Bruecker parameter rs for the density and spatial dimension d the differences for d ∈ {2,3} reflect the different expressions for the Coulomb interaction, whichs Fourier component is vq = rac{4π}{q^2} in 3D and vq = rac{2π}{q} in 2D the one-dimensional case involves a short-distance cutoff to avoid the non-integrable divergence of the interaction at zero distance and is not implemented